Fix:graphics_qt_qpainter:Fixed some 64 bit problems

+++ navit/graphics/qt_qpainter/graphics_qt_qpainter.cpp
@@ -1037,12 +1037,12 @@ struct event_watch {
 static struct event_watch *
 event_qt_add_watch(void *fd, enum event_watch_cond cond, struct callback *cb)
 {
-	dbg(1,"enter fd=%d\n",(int)fd);
+	dbg(1,"enter fd=%d\n",(int)(long)fd);
 	struct event_watch *ret=g_new0(struct event_watch, 1);
 	ret->fd=fd;
 	ret->cb=cb;
 	g_hash_table_insert(event_gr->widget->watches, fd, ret);
-	ret->sn=new QSocketNotifier((int) fd, QSocketNotifier::Read, event_gr->widget);
+	ret->sn=new QSocketNotifier((int)(long)fd, QSocketNotifier::Read, event_gr->widget);
 	QObject::connect(ret->sn, SIGNAL(activated(int)), event_gr->widget, SLOT(watchEvent(int)));
 	return ret;
 }
@@ -1069,7 +1069,7 @@ event_qt_add_timeout(int timeout, int multi, struct callback *cb)
 static void
 event_qt_remove_timeout(struct event_timeout *ev)
 {
-	event_gr->widget->killTimer((int) ev);
+	event_gr->widget->killTimer((int)(long)ev);
 	g_hash_table_remove(event_gr->widget->timer_callback, ev);
 	g_hash_table_remove(event_gr->widget->timer_type, ev);
 }
